+++ ali/common/src/main/java/com/yanny/ali/plugin/common/NodeUtils.java
@@ -197,7 +197,9 @@ public class NodeUtils {
     public static boolean hasPredicates(IServerUtils utils, List<LootItemCondition> conditions) {
         List<ResourceLocation> ignored = utils.getConfiguration().ignoredPredicateConditions;
 
-        return conditions.stream().anyMatch((c) -> !ignored.contains(BuiltInRegistries.LOOT_CONDITION_TYPE.getKey(c.getType())));
+        return conditions.stream()
+                .flatMap((c) -> utils.unwrapCondition(utils, c).stream())
+                .anyMatch((c) -> !ignored.contains(BuiltInRegistries.LOOT_CONDITION_TYPE.getKey(c.getType())));
     }
 
     @NotNull