+++ ali/common/src/main/java/com/yanny/ali/plugin/server/TooltipUtils.java
@@ -25,13 +25,12 @@ import net.minecraft.network.chat.contents.TranslatableContents;
 import net.minecraft.world.level.storage.loot.entries.LootPoolSingletonContainer;
 import net.minecraft.world.level.storage.loot.entries.NestedLootTable;
 import net.minecraft.world.level.storage.loot.functions.*;
-import net.minecraft.world.level.storage.loot.predicates.BonusLevelTableCondition;
-import net.minecraft.world.level.storage.loot.predicates.LootItemCondition;
-import net.minecraft.world.level.storage.loot.predicates.LootItemRandomChanceCondition;
-import net.minecraft.world.level.storage.loot.predicates.LootItemRandomChanceWithEnchantedBonusCondition;
+import net.minecraft.world.level.storage.loot.predicates.*;
 import net.minecraft.world.level.storage.loot.providers.number.ConstantValue;
 import net.minecraft.world.level.storage.loot.providers.number.NumberProviders;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+import org.jetbrains.annotations.Unmodifiable;
 
 import java.io.Serializable;
 import java.lang.reflect.Field;
@@ -132,6 +131,36 @@ public class TooltipUtils {
      */
     public static boolean isConditional(LootItemConditionalFunction function) {
         return !function.predicates.isEmpty();
+    }
+
+    @NotNull
+    public static List<LootItemCondition> unwrapAllOf(IServerUtils ignoredUtils, AllOfCondition condition) {
+        return condition.terms;
+    }
+
+    @Unmodifiable
+    @Nullable
+    public static List<LootItemCondition> unwrapConditionReference(IServerUtils utils, ConditionReference condition) {
+        return utils.getServerLevel().getServer().reloadableRegistries().lookup().get(Registries.PREDICATE, condition.name())
+                .map((holder) -> Collections.singletonList(holder.value()))
+                .orElse(null);
+    }
+
+    @NotNull
+    public static List<LootItemFunction> unwrapFunctionSequence(IServerUtils ignoredUtils, SequenceFunction function) {
+        return function.functions;
+    }
+
+    @Unmodifiable
+    @Nullable
+    public static List<LootItemFunction> unwrapFunctionReference(IServerUtils utils, FunctionReference function) {
+        if (isConditional(function)) {
+            return null;
+        }
+
+        return utils.getServerLevel().getServer().reloadableRegistries().lookup().get(Registries.ITEM_MODIFIER, function.name)
+                .map((holder) -> Collections.singletonList(holder.value()))
+                .orElse(null);
     }
 
     @NotNull